+++ src/assets/redactorplugin/RedactorPluginAsset.php
@@ -4,6 +4,7 @@ namespace carlcs\redactorcustomstyles\assets\redactorplugin;
 
 use carlcs\redactorcustomstyles\Plugin;
 use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
 
 class RedactorPluginAsset extends AssetBundle
 {
@@ -12,12 +13,17 @@ class RedactorPluginAsset extends AssetBundle
      */
     public function init()
     {
-        $this->sourcePath = __DIR__.'/dist/assetbundle';
+        $this->sourcePath = __DIR__.'/dist';
+
+        $this->depends = [
+            CpAsset::class,
+        ];
 
         if (Plugin::getInstance()->getSettings()->svgPolyfill) {
             $this->js[] = 'svg-polyfill.js';
         }
 
+        $this->js[] = 'customstyles.js';
         $this->css[] = 'customstyles.css';
 
         parent::init();